Update sdReq with user input
