+++ lib/actions/jlinkTargetActions.js
@@ -37,9 +37,9 @@
 import { logger } from 'nrfconnect/core';
 import nrfjprog from 'pc-nrfjprog-js';
 import MemoryMap from 'nrf-intel-hex';
-import { checkUpToDateFiles } from './fileActions';
 import { getMemoryRegions, getTargetRegions } from '../util/regions';
 import * as targetActions from './targetActions';
+import * as fileActions from '../actions/fileActions';
 import * as devices from '../util/devices';
 
 // Get device infos by calling nrfjprog
@@ -214,7 +214,7 @@ export function updateTargetRegions(regionsInput, memMap, deviceInfo) {
         const regions = getTargetRegions(memMaps, loadedDevice, deviceInfo);
 
         dispatch(targetActions.targetRegionsKnownAction(regions));
-        dispatch(targetActions.updateFileAppRegions());
+        dispatch(fileActions.updateFileAppRegions());
     };
 }
 
@@ -338,7 +338,7 @@ export function write() {
         // FIXME: Check if the target's UICR is blank. If not, slice the flattened
         // hex files so that the code doesn't try to overwrite UICR.
         // This is part of the «UICR can only be written to after an "erase all"» logic
-        return checkUpToDateFiles(dispatch, getState).then(() => {
+        return fileActions.checkUpToDateFiles(dispatch, getState).then(() => {
             let pages = MemoryMap.flattenOverlaps(
                 MemoryMap.overlapMemoryMaps(appState.file.memMaps),
             ).paginate(pageSize);
