+++ lib/actions/targetActions.js
@@ -38,8 +38,6 @@ import { List } from 'immutable';
 import * as jlinkTargetActions from './jlinkTargetActions';
 import * as usbsdfuTargetActions from './usbsdfuTargetActions';
 import { CommunicationType } from '../util/devices';
-import { RegionName } from '../util/regions';
-import { fileRegionsKnownAction } from './fileActions';
 
 export const TARGET_TYPE_KNOWN = 'TARGET_TYPE_KNOWN';
 export const TARGET_INFO_KNOWN = 'TARGET_INFO_KNOWN';
@@ -149,51 +147,6 @@ export function userInputUpdated(userInput) {
     return {
         type: USER_INPUT_UPDATED,
         userInput,
-    };
-}
-
-// There is an Application on top of SoftDevice in the .hex file,
-// but there is no SoftDevice in the .hex file,
-// In this case, if there is a SoftDevice being found in target device,
-// then the Application region should be displayed.
-// If there is no SoftDevice in both .hex file and target device,
-// then the user should give input instead.
-// (Or fix getting softdevice id from bootloader)
-export function updateFileAppRegions() {
-    return (dispatch, getState) => {
-        let fileRegions = getState().app.file.regions;
-        const targetRegions = getState().app.target.regions;
-        const deviceInfo = getState().app.target.deviceInfo;
-
-        // Assume that the region on top of the SoftDevice is application.
-        // Assume also that the region on top of the MBR which is not SoftDevice is application.
-        const fileSoftDeviceRegion = fileRegions.find(r => r.name === RegionName.SOFTDEVICE);
-        const targetSoftDeviceRegion = targetRegions.find(r => r.name === RegionName.SOFTDEVICE);
-        const pageSize = deviceInfo.pageSize;
-        if (!fileSoftDeviceRegion && targetSoftDeviceRegion) {
-            const softDeviceEnd = targetSoftDeviceRegion.startAddress
-                + targetSoftDeviceRegion.regionSize;
-            let appRegion = fileRegions.find(r =>
-                r.startAddress === Math.ceil((softDeviceEnd) / pageSize) * pageSize);
-            if (appRegion) {
-                const appRegionIndex = fileRegions.indexOf(appRegion);
-                appRegion = appRegion.set('name', RegionName.APPLICATION);
-                fileRegions = fileRegions.set(appRegionIndex, appRegion);
-                dispatch(fileRegionsKnownAction(fileRegions));
-            }
-        }
-
-        // Remove Application label if there is no SoftDevice region existing.
-        // TODO: fix after dfu
-        if (!fileSoftDeviceRegion && !targetSoftDeviceRegion) {
-            let appRegion = fileRegions.find(r => r.name === RegionName.APPLICATION);
-            if (appRegion) {
-                const appRegionIndex = fileRegions.indexOf(appRegion);
-                appRegion = appRegion.set('name', RegionName.NONE);
-                fileRegions = fileRegions.set(appRegionIndex, appRegion);
-                dispatch(fileRegionsKnownAction(fileRegions));
-            }
-        }
     };
 }
