+++ lib/actions/usbsdfuTargetActions.js
@@ -41,6 +41,7 @@ import SerialPort from 'serialport';
 import * as dfujs from 'pc-nrf-dfu-js';
 import { initPacket, detachAndWaitFor, waitForDevice, dfuTrigger } from 'nrf-device-setup';
 import * as targetActions from './targetActions';
+import * as fileActions from '../actions/fileActions';
 import * as devices from '../util/devices';
 import { Region, RegionPermission, RegionName, getSoftDeviceId } from '../util/regions';
 import { setExpectedSerialNumber } from './deviceChangeActions';
@@ -156,7 +157,7 @@ function loadDeviceInfo(selectedDevice) {
             });
 
             dispatch(targetActions.targetRegionsKnownAction(regions));
-            dispatch(targetActions.updateFileAppRegions());
+            dispatch(fileActions.updateFileAppRegions());
             dispatch(targetActions.updateTargetWritable());
         } catch (versionError) {
             logger.error(`Error when fetching device versions: ${versionError}`);
@@ -227,13 +228,6 @@ function createDfuImages() {
     };
 }
 
-export function updateUserInput(filed, value) {
-    return (dispatch, getState) => {
-        const userInput = getState().app.target.userInput;
-        dispatch(targetActions.userInputUpdated(userInput.set(filed, value)));
-    };
-}
-
 // Check if the files can be written to the target device
 export function canWrite() {
     return (dispatch, getState) => {
@@ -249,13 +243,22 @@ export function canWrite() {
         }
         const fileMemMap = getState().app.file.memMaps[0][1];
         const deviceInfo = getState().app.target.deviceInfo;
-        if (!getSoftDeviceId(fileMemMap, deviceInfo)) {
+        const userInput = getState().app.target.userInput;
+        if (!getSoftDeviceId(fileMemMap, deviceInfo) && !userInput.appSdReq.length) {
             dispatch(targetActions.targetWarningKnownAction('User input is needed.'));
             dispatch(targetActions.userInputEnabled(['appSdReq']));
             return;
         }
 
         dispatch(targetActions.targetWritableKnownAction(true));
+    };
+}
+
+export function updateUserInput(filed, value) {
+    return (dispatch, getState) => {
+        const userInput = getState().app.target.userInput;
+        dispatch(targetActions.userInputUpdated(userInput.set(filed, value)));
+        dispatch(canWrite());
     };
 }
 
@@ -334,13 +337,16 @@ function handleUserInput(imageInput, userInput) {
         fwVersion = userInput.sdVer;
     } else if (image.name.startsWith(RegionName.APPLICATION)) {
         fwVersion = userInput.appVer;
+        // Handle sdReq
+        if (!image.initPacket.sdReq) {
+            if (!userInput.appSdReq) {
+                throw new Error('User input is needed but it is not set.');
+            }
+            sdReq = [parseInt(userInput.appSdReq, 16)];
+            console.log('sdReq', sdReq);
+        }
     } else {
         throw new Error('Firmware type is unknown when setting user input.');
-    }
-
-    // Handle sdReq
-    if (!image.initPacket.sdReq) {
-
     }
 
     image = sdReq ? image.setIn(['initPacket', 'sdReq'], sdReq) : image;
@@ -394,6 +400,8 @@ function operateDFU(inputDfuImages, inputDevice) {
         .then(() => {
             serialNumber = getState().app.target.serialNumber;
             image = dfuImages.first();
+            console.log('lalalalal');
+            console.log(image);
             const packet = initPacket.createInitPacketUint8Array(image.initPacket);
             const firmwareUpdates = new dfujs.DfuUpdates([{
                 initPacket: packet,
@@ -443,6 +451,7 @@ export function write() {
         const userInput = getState().app.target.userInput;
         let dfuImages = getState().app.target.dfuImages;
 
+        console.log(dfuImages);
         dfuImages = dfuImages.map(image =>
             handleImage(image, fileRegions, fileMemMap));
         dfuImages = dfuImages.map(image =>
