+++ lib/reducers/targetReducer.js
@@ -131,6 +131,9 @@ export default function target(state = new InitialState(), action) {
                 .set('writtenAddress', 0)
                 .set('isReady', true);
 
+        case targetActions.USER_INPUT_ENABLED:
+            return state.set('userInputFields', action.fields);
+
         case targetActions.USER_INPUT_UPDATED:
             return state.set('userInput', action.userInput);
